+++ src/FirstComponent/FirsrComponent.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-
+import '../App.css';
 const FirstComponent = () => {
 	return (
-		<div>
+		<div className="component first">
 			<h1>Плачу на техно</h1>
 		</div>
 	);
