+++ src/index.css
@@ -1,7 +1,7 @@
 body {
 	margin: 0;
 	padding: 0;
-	height: 100%;
+
 	overflow: hidden;
 	font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
 		'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
@@ -11,7 +11,6 @@ body {
 }
 
 code {
-	overflow: auto;
 	font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
 		monospace;
 }
